tidy(nodes): bring masked blend latents masking logic into invoke core
